pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["obsidian_vault_mcp"]

[project]
name = "obsidian-vault-mcp"
version = "0.1.0"
description = "Obsidian vault MCP server — read, write, and search vault files directly on disk"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
]
